getA {pedigreemm}	R Documentation

Additive Relationship Matrix

Description

Additive relationship matrix from a pedigree

Usage

getA(ped)

Arguments

ped
a pedigree that includes the individuals who occur in labs

Details

Returns the additive relationship matrix for the pedigree ped .

Value

Sparse matrix

Examples

## Example from chapter 2 of Mrode (2005)

ped <- pedigree(sire = c(NA,NA,1, 1,4,5),
                dam  = c(NA,NA,2,NA,3,2), label= 1:6)
(getA(ped))

[Package pedigreemm version 0.3-4 Index]