mastitis {pedigreemm}	R Documentation

Mastitis cases in dairy cattle

Description

Records of the number of cases of clinical mastitis during the first lactation of 1,675 cows, primarily Holsteins. Cows belonged to 41 herds and were daughters of 38 sires. There were 1,491 healthy cows, 134 had only one case of mastitis, 36 had 2 cases, and 14 had between 4 and cases. Overall, mastitis incidence was 0.11. Calving years for these records were from 2000 through 2005. The sire, herd and days in milk are also recorded for each cow.

Format

A data frame with 1675 observations on the following 8 variables.

id

Identifier of the animal.

sire

Identifier of the animal's sire.

birth

year of birth of the animal (as a factor).

herd

herd id number (as a factor).

calvingYear

year of calving for this lactation.

DIM

total number of days in milk for the lactation.

mastitis

a factor indicating if the cow had any incidents of clinical mastitis during the lactation.

NCM

An ordered factor giving the number of clinical mastitis cases for the cow during this lactation.

Details

The pedigree of the sires is given in the companion pedSires data set.

Source

Vazquez, A.I. 2007. Analysis of number of episodes of clinical mastitis in Norwegian Red and Holstein cows with Poisson and categorical data mixed models. Master of Science Thesis. University of Wisconsin - Madison. 162 pp.

2010. A.I. Vazquez, D.M. Bates, G.J.M. Rosa, D. Gianola and K.A. Weigel. Technical Note: An R package for fitting generalized linear mixed models in animal breeding. Journal of Animal Science, 88:497-504.

References

2010. A.I. Vazquez, D.M. Bates, G.J.M. Rosa, D. Gianola and K.A. Weigel. Technical Note: An R package for fitting generalized linear mixed models in animal breeding. Journal of Animal Science, 88:497-504.

See Also

pedSires, pedigree

Examples

str(mastitis)
summary(mastitis, maxsum = 10)

[Package pedigreemm version 0.3-4 Index]