pedCowsR {pedigreemm}	R Documentation

Pedigree of the cows in milk with 0.70 of the information in pedCows

Description

A pedigree object giving (part of) the pedigree of the cows in the milk data frame. This pedigree allows the example with 'milk' to run faster.

Format

The format is: Formal class 'pedigree' [package "pedigreemm"] with 3 slots ..@ sire : int [1:6547] NA NA NA NA NA NA NA NA NA NA ... ..@ dam : int [1:6547] NA NA NA NA NA NA NA NA NA NA ... ..@ label: chr [1:6547] "1" "2" "3" "4" ...

References

2010. A.I. Vazquez, D.M. Bates, G.J.M. Rosa, D. Gianola and K.A. Weigel. Technical Note: An R package for fitting generalized linear mixed models in animal breeding. Journal of Animal Science, 88:497-504.

See Also

milk

Examples

str(pedCowsR)

[Package pedigreemm version 0.3-4 Index]