pedigree {pedigreemm}	R Documentation

Pedigree Constructor

Description

Construct an object of class "pedigree", more conveniently than by new("pedigree", ....).

Usage

pedigree(sire, dam, label)

Arguments

sire
numeric vector (with some NA entries) of integer IDs, denoting a previous entry in the pedigree corresponding to the current entry's “father”.

dam
similarly as sire for the “mother” of each entry.

label
a vector coercable to "character" of the same length as sire and dam giving a unique ID for the corresponding entry.

Value

an object of formal class "pedigree".

References

2010. A.I. Vazquez, D.M. Bates, G.J.M. Rosa, D. Gianola and K.A. Weigel. Technical Note: An R package for fitting generalized linear mixed models in animal breeding. Journal of Animal Science, 88:497-504.

See Also

the pedigree class.

Examples

example("pedigree-class") ## 'p1' pedigree object `the hard way'

ped <- pedigree(sire = c(NA,NA,1, 1,4,5),
                dam  = c(NA,NA,2,NA,3,2), label= 1:6)
## note that 'label' is coerced to character automatically
ped
stopifnot(identical(ped, p1))

[Package pedigreemm version 0.3-4 Index]