pedigree-class {pedigreemm}	R Documentation

Class "pedigree"

Description

Objects of class "pedigree" represent a set of individuals that can have two parents including their parent-child relations. The terminology has been taken from cattle breeding. The "pedinbred" class is an extension of the pedigree class with an additional slot of the inbreeding coefficients.

Objects from the Class

Objects in the "pedigree" class can be created by calls of the form new("pedigree", ...), or more conveniently, pedigree(sire= ., dam = ., label =.).

Objects of the "pedinbred" class are created by coercing a pedigree to class "pedinbred".

Slots

sire:

integer vector (with some NA entries), denoting a previous entry in the pedigree corresponding to the current entry's “father”.

dam:

similarly as sire for the “mother” of each entry.

label:

a "character" vector of the same length as sire and dam giving a unique ID for the corresponding entry.

F:

(class "pedinbred" only) a numeric vector of inbreeding coefficients.

Methods

coerce

signature(from = "pedigree", to = "sparseMatrix"): returns a sparse, unit lower-triangular matrix which is the inverse of the "L" part of the "LDL'" form of the Cholesky factorization of the relationship matrix. All non-zero elements below the diagonal are -0.5.

coerce

signature(from = "pedigree", to = "data.frame"): ...

head

signature(x = "pedigree"): ...

show

signature(object = "pedigree"): ...

tail

signature(x = "pedigree"): ...

References

R. A. Mrode, Linear Models for the Prediction of Animal Breeding Values, 2nd ed, CABI Publishing, 2005.

2010. A.I. Vazquez, D.M. Bates, G.J.M. Rosa, D. Gianola and K.A. Weigel. Technical Note: An R package for fitting generalized linear mixed models in animal breeding. Journal of Animal Science, 88:497-504.

See Also

pedigree, inbreeding

Examples

## Rather use, pedigree()!   The following is "raw code":
## Example from chapter 2 of Mrode (2005)
p1 <- new("pedigree",
          sire = as.integer(c(NA,NA,1, 1,4,5)),
          dam  = as.integer(c(NA,NA,2,NA,3,2)),
          label = as.character(1:6))
p1
(dtc <- as(p1, "sparseMatrix"))   # T-inverse in Mrode's notation
solve(dtc)
inbreeding(p1)

[Package pedigreemm version 0.3-4 Index]