pedigreemm-class {pedigreemm}	R Documentation

Pedigree-based mixed-effects model fits

Description

A mixed-effects model fit by pedigreemm. This class extends class "merMod" class and includes one additional slot, relfac, which is a list of (left) Cholesky factors of the relationship matrices derived from "pedigree" objects.

Objects from the Class

Objects are created by calls to the pedigreemm function.

Slots

relfac:

A list of relationship matrix factors. All other slots are inherited from class "merMod".

Extends

Class "merMod", directly.

Methods

fitted

signature(object = "pedigreemm"): actually a non-method in that fitted doesn't apply to such objects because of the pre-whitening.

ranef

signature(object = "pedigreemm"): incorporates the pedigree into the random effects as returned for the object viewed as a "merMod)" object.

residuals

signature(object = "pedigreemm"): also a non-method for the same reason as fitted

See Also

pedigreemm

Examples

showClass("pedigreemm")

[Package pedigreemm version 0.3-4 Index]